+++ Source/Beezer/BeezerListView.cpp
@@ -80,7 +80,7 @@ BeezerListView::~BeezerListView()
     // Delete all list items here as CLV doesn't do it for us in its destructor
     int32 itemCount = FullListCountItems();
     for (int32 i = 0; i < itemCount; i++)
-        delete(ListEntry*)ColumnListView::RemoveItem(0);
+        delete(ListEntry*)ColumnListView::RemoveItem((int32)0);
 }
